Fixed broken versions of MIN and MAX macro
The previous definition of MIN(A,B) and MAX(A,B) generates invalid
expressions if the result is part of a larger expression. E.g.
MIN(a,b) / c would evaluate to ((a) < (b)) ? (a) : (b) / c. Because of
operator precedence, this will not generate the expected result.
Additional parentheses were added to avoid these types of errors.

This bug was discovered in fix_wall_gran_hooke.cpp. All other
occurrences of the bad macro definition are now patched as well.