fix issue with atom types
We need to set the min/max atom types we are going to use right at the start,
because LIGGGHTS scans all fixes for the min/max atom types they use in advance
and allocates only those in the found range and not all that you may have
defined via create_box and fix property/global!
Since in this case we don't know in advance which atom types we will insert,
we assume that we will use all available types (1 to atom->ntypes).