add option for synchronous mpi send
with MPI_Send it is possible that the data ends up in a message
buffer until the matching MPI_Recv occurs. This buffer may grow
rather large if the sending and receiving processes have very
different runtimes. To avoid this memory overhead one can choose
to use the synchronous MPI_Ssend which blocks until the sent data
has been received.