base AssociativePointerArray on std::map
For proper restart of per element mesh properties read from VTK
files, it is necessary to keep the list of properties in a
well-defined order. This was not the case when a property (here
"lineNo") was removed. Moving the last array element to the free
slot was destroying the ordering.