add test case to analyze contact force models
Follows the analysis of rebound kinematics in
Thornton et al., Powder Technol. 210 (2011) 189-197
(elastic collisions) and
Thornton et al., Powder Technol. 233 (2013) 30-46
(inelastic collisions)
Simulates the oblique impact of a sphere with a target wall. The sphere
has an inital velocity of 5 m/s, no initial angular velocity and an
impact angle of 1°-45°
The simulation outputs the normalized rebound surface velocity,
normalized rebound angular velocity and tangential coefficient of
restitution as a function of the normalized impact angle.
In case of elastic impact and an impact angle of 5°, the tangential
force is saved together with the tangential displacement (run
processData.m)
Note that in contrast to Thornton et al. LIGGGHTS is equating maximum
normal overlap instead of collision time to get the normal stiffness for
the hooke/LSDP model.