Solve merge conflicts for fix_chem_shrink_core
- Change k0 and Ea values from constants to arrays (these values are
read from input script and subject to change per simulation)
- Same with the layerDensity.
- rmin was changed to rrmin (minimum relative radius) - this was set
back again to rmin (doesn't affect the calculation or the procedure at
all).
- Get Aterm is not calculated seperately for reacting gas. The arrays
for fix_k0_ and fix_Ea_, are directly used, and they change because of
reactant gas in the input script provided.
- MassT calculation had been changed, but while merging reverted to
merge version.
- instead of nevery, scale_reduction_rate is used. Thus added as a fix.
This value is written in the fix, and used from there.
- In merge file, calculation of new mass layers were carried out
wrongly, ( <= layers) should have been used. (This was changed already
in the local branch but not pushed).
- Re-calculation of density, porosity fields is reverted to merge
version (it was altered as porosity and density was switched to time
based). After testing merge, re-add and verify again.
- Fix nameing for Low Reduction Fractional Reduction.
- equal sign for delete fix_Ea_.
- layerDensities is per-particle so particle counter is added.