fix creation of bond history
The issue here is that at startup, LIGGGHTS/LAMMPS immediately creates
an atom vector of length 1, no matter if there are any particles or not.
At this point, however, nothing in the input script has been parsed,
especially not the bond_style command which sets the size of the bond
history in the Atom class and n_bondhist still has its default value 0.
Thus there exists an atom vector which has no bond history array and as
soon as a first particle is moved (via exchange) to an initially empty
partition an access violation occurs.
The proper solution should be to allocate the bond history array when
setting the size of the bond history.