improve search tree generation for region mesh/tet
if we're unlucky all nodes of a tetrahedron are on the edges of the
bounding box that are not considered to be *inside*, and we miss to
insert that cell into the search tree
as a minor improvement, we additionally check against the tetrahedron's
center now, but ultimately this would need a full bounding box -
tetrahedron intersection test (e.g. H. Ratschek and J. Rokne: Intern. J.
Computer Math., vol 65, 191-204, 1997)