fix hybrid atom style with varying particle radius
if atom style 'hybrid' had a sub-style 'sphere' (i.e. 'granular') or
'sph/var' in combination with a fix that modifies the atom radius (e.g.
fix adapt, fix remove), an incorrect size_forward was used; the reason
for this is that atom style 'hybrid' calculates the sizes in the
'settings' method, while 'sphere' and 'sph/var' adjust the size in their
init methods at a later point, resulting in a size mismatch;

resolve this by recalculating size_forward (if necessary) after all
sub-styles have called their init methods;

note that the current LAMMPS version solves this situation differently:
there the sub-styles have an additional setting option if one whishes to
modify the radius during the simulation; thus all size adjustments are
done in the 'settings' methods and none in the 'init' methods